Forgot to move headers to Include directory for Box2D

+++ Harmony/include/Box2D/Common/b2Timer.h
@@ -19,7 +19,7 @@
 #ifndef B2_TIMER_H
 #define B2_TIMER_H
 
-#include <Box2D/Common/b2Settings.h>
+#include "Box2D/Common/b2Settings.h"
 
 /// Timer for profiling. This has platform specific code and may
 /// not work on every platform.
@@ -42,8 +42,8 @@ private:
 	float64 m_start;
 	static float64 s_invFrequency;
 #elif defined(__linux__) || defined (__APPLE__)
-	unsigned long m_start_sec;
-	unsigned long m_start_usec;
+	unsigned long long m_start_sec;
+	unsigned long long m_start_usec;
 #endif
 };
